+++ packages/core/index.tsx
@@ -4,6 +4,7 @@ import { createLoggingService } from './externals/logging'
 import { createGlobalData } from './externals/pixivGlobalData'
 import { createAPIClient } from './externals/apiClient'
 import { createAddonStore } from './externals/addonStore'
+import { ServiceProvider } from './components/ServiceProvider'
 import { App } from './components/App'
 
 const loggingService = createLoggingService()
@@ -12,11 +13,13 @@ const apiClient = createAPIClient(globalData, loggingService)
 const addonStore = createAddonStore()
 
 ReactDOM.render(
-  <App
+  <ServiceProvider
     loggingService={loggingService}
     apiCllient={apiClient}
     addonStore={addonStore}
-  />,
+  >
+    <App />
+  </ServiceProvider>,
   document.body.appendChild(document.createElement('div'))
 )
